# Module-2/02-stu-record-store-api-doc/starter/record-store-api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /records/{id}:
    get:
      tags:
        - record-store-controller
      operationId: getRecordById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Record'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    put:
      tags:
        - record-store-controller
      operationId: updateRecordById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Record'
        required: true
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    delete:
      tags:
        - record-store-controller
      operationId: deleteRecordById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
  /records:
    get:
      tags:
        - record-store-controller
      operationId: getAllRecords
      parameters:
        - name: artist
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Record'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
    post:
      tags:
        - record-store-controller
      operationId: createRecord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Record'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Record'
        '422':
          description: Unprocessable Entity
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomErrorResponse'
components:
  schemas:
    CustomErrorResponse:
      type: object
      properties:
        errorMsg:
          type: string
        status:
          type: integer
          format: int32
        errorCode:
          type: string
        timestamp:
          type: string
          format: date-time
    Record:
      required:
        - album
        - artist
        - year
      type: object
      properties:
        artist:
          type: string
        album:
          type: string
        year:
          maxLength: 4
          minLength: 4
          type: string
        id:
          type: integer
          format: int32